cloudmusic/requirements.txt
Django==2.1.7
mysqlclient==1.4.1
wheel==0.32.3
gunicorn==19.9.0
psycopg2==2.7.7
requests == 2.21.0
google-auth == 1.6.3